fixed the input problem (interface issue)
the problem was that interfaces are not allowed to have fields that are
not public static final, which means that things like the "name" field
for extractors were not able to override the default empty string. I
think the solution is to just make sure that each extractor has a static
field for name, but this can't really be enforced by the compiler.