better handling of unaligned words
Some rules were not being extracted because they can't be concatenated
out of tightly-bound phrases on the source side. The Hiero specification
only says that the elided NTs must be tight bound, which means there may
be some contiguous stretches of terminals that have unaligned words on
the boundary. This change allows that in the extraction.