extraction consistent with Dyer's implementation
Modification: instead of appending a whole phrase's worth of terminals
at once, extend the rule only one terminal symbol at a time. Use the
phrases for extending with an NT. As of this commit, running
GrammarComparison against cdec's extraction output for 1 sentence of
es-en europarl shows that the grammars are identical.