serious modification to command line options
This makes it easier to pass the configuration options around because
they're actually parsed in ThraxConfig rather than just sitting as
Strings until someone actually needs the result. Also, I have a sneaking
suspicion that I broke the SAMT somehow, because it's not producing any
arity-2 rules, so we'd better look into that.