refactoring of NT labelling for rules
getLabelVariants has its main implementation in Hiero now, and SAMT
defers to that. The only difference, then, is providing a different set
of labels for each span.

Also, the hashCode problem for rules has been fixed. I'm not even sure
what the problem was, but for some reason the yield for the rule wasn't
being recalculated when it was supposed to, resulting in almost all
rules having an empty yield when considering their hashcode. But now it
seems to work? I thought the original problem was the memoizing of the
yield calculation, because when I removed it, the problem went away. But
when I put it back in, there was still no problem. It's very weird.