fixed loose phrase boundary implementation
The problem was that the extraction of loose phrases allowed a phrase
with no aligned words on the source side, which doesn't make any sense.
I think there's a better way to do this, which may be included in the
future, but this at least avoids index out of bounds errors for now.