major rule scoring modification
1) removed score array from Rule data structure entirely -- it's not
actually relevant to the usage of Rule

2) instead of attaching features to the extractor or leaving them in the
main method, they are attached to a special scorer object, which
incrementally keeps track of the scores for each rule

3) by incrementally, we mean update it when necessary because we've seen
a rule another time. There are three ways to do this: don't update
(because the score can't possible be different instance-to-instance,
e.g. in relative frequency) or take the maximum or minimum of the
current score or the new rule's score. For example, in lexprob different
instances may have different scores because they have different
underlying alignments. We want to take the minimum (negative log prob).