java-native multithreading
The main Thrax method uses a thread pool to process many input sentences
at once. We had to change the noteExtraction methods in most features to
make sure that they were synchronized. This usually was done by adding a
"synchronized" modifier to the method declaration. But for the
LexicalProbabilityFeature, we synchronized the underlying
CollocationTable instead.