main methods for the extractors
In case you want to just see what rules get extracted, without worrying
about all of the feature stuff and hadoop things. The main point was to
do some extraction profiling with jrat, but there wasn't anything that
obviously needed to be tuned, so that's kind of a dead end for the
moment.