removed all normalization
We were running into annoying special cases, and we would have to keep
track and make sure that we normalize in the same way every time we look
at the same input. This way, the obvious behavior (not to change the
input tokens without telling the user) is preserved. Not to mention, it
often helps to have -lrb-, -rrb- etc in the output because we might
compare to a normalized test file in some way.

The main takeaway is, DON'T CHANGE THE USER'S INPUT. If they want to
normalize a particular way, they can do it themselves.