new filtering strategy
For each word in the test set, we keep a set of sentence numbers that
contain that word. Then, when we want to test a rule to see if it's in
the test set, we only have to compare against the intersection of these
sets, because the sentences in the intersection are the only possible
sentences that might match the rule.