HieroLabeler uses singleton instead of empty set
The way the hierarchical extractor is written right now, having an empty
label set is different from having the default label. However, this
commit is buggy: it assigns labels to some spans that shouldn't be
assigned any. This is weird, because I was under the impression that
every non-terminal in a rule has to be pulled from the phrase pairs and
therefore has to have been assigned at least some kind of label.
Hopefully this doesn't indicate a bug in the master branch.