equivalent output to master on example sentences
The problem was that the span of rules is in fact limited to the initial
length limit, not to the so-called rule span limit. This is probably a
bug in the master branch. It doesn't affect SAMT, since those values are
equal by default, but it does affect Hiero.

This branch now has the same limit, but it's not clear whether it
should. The reason the limit showed up in master was that master refuses
to extract rules if they don't have a label, and the only phrase pairs
that have labels are the initial phrase pairs calculated for each
sentence. Therefore, if the RHS of a rule doesn't coincide with one of
those phrase pairs, you're sunk. That's why a rule can only have a span
the size of an initial phrase pair.

I've manually added that limit here, but we might want to rework it
during this rewrite.