bugfix in fast filtering
We split the source side of the rule string on nonterminals, then make sure that
each ngram in the rule was found in the test set, short-circuiting on the first such
n-gram in the rule not to be found.  The bug was that we were quitting on the empty
string, which is found any time a nonterminal starts or ends a source side, causing
us to erroneously reject all such rules.

An alternative fix would have been to have put the empty string in the list of n-grams.