Only use _S label for rules with long LHS spans
For very short training pairs, like dictionary entries, we don't
want to assign the (rather uninformative) label _S when something
better like a POS is available.

We reserve the special "_S" symbol for rules that could only be
extracted because the rule span limit is lifted for rules that
span the whole sentence.