Fix rule length limits for extracted rules
Case 1: hierarchical rules
In this case, extracted rules are not allowed to have more than
nonlex-{source,target}-length symbols on the right hand side.

Case 2: lexical rules
Here, the rules are restricted to the length of the initial phrase
pair. Initially, that happened automatically due to the way rules
are extracted, but now that the full sentence is a possible LHS,
we need to check for it explicitly.

NOTE: we're about to introduce a new configuration parameter that
will allow restricting the length of lexical rules further.