Bugfix related to lexical probability features + correct exit code on failure + use number of reducers from hadoop, when set.
Details:
TrieLexprobTable consumes the output of WordLexprobJob. It expects the output to be sorted in order to use binary search later. So far this was not an issue as the default setting for the number of reducers is 1, which led to a sorted output. If the number of reducers is increased this fails however. The solution is to partition the output of WordLexprobJob is sorted.