add source and target raw count features
THIS UNTESTED. I didn't even try to compile it because this computer
doesn't have ant for some reason.

This commit adds features that emit the raw counts for the source and
target side of a given rule. That is, for a rule A -> (f,e) these
features return the number of times f was seen as a source side (summed
over the entire corpus), and similarly for e.

It's a bad hack, basically just a cut-and-paste job the phrasal
probability features, which already had to calculate this sum as the
denominator of their probabilities.